+++ sherpa/cpp_api/CMakeLists.txt
@@ -2,7 +2,7 @@ set(sherpa_cpp_api_srcs
   feature-config.cc
   offline-stream.cc
   offline-recognizer.cc
-  online_recognizer.cc
+  online-recognizer.cc
 )
 add_library(sherpa_cpp_api ${sherpa_cpp_api_srcs})
 target_link_libraries(sherpa_cpp_api sherpa_core)
