+++ sherpa/cpp_api/online-stream.h
