+++ sherpa/csrc/test-hypothesis.cc
